Divide test classes by feature
