+++ src/test/java/api/actions/CardActions.java
@@ -11,29 +11,27 @@ import static io.restassured.RestAssured.given;
 
 public class CardActions {
 
-    public Card createCard(String name, String idList){
+    public Card createCard(String name, String idList) {
         RestAssured.baseURI = "https://api.trello.com/1";
         RequestSpecification requestSpecification = given()
                 .contentType(ContentType.TEXT)
                 .queryParam("name", name)
                 .queryParam("idList", idList)
-                .queryParam("keepFromSource","all")
+                .queryParam("keepFromSource", "all")
                 .queryParam("key", System.getenv("trl_key"))
                 .queryParam("token", System.getenv("trl_token"));
         return requestSpecification.request(Method.POST, "/cards/").as(Card.class);
     }
 
-    public Checklist createChecklist(String idCard, String name){
+    public Checklist createChecklist(String name, String idCard) {
         RestAssured.baseURI = "https://api.trello.com/1";
         RequestSpecification requestSpecification = given()
                 .contentType(ContentType.TEXT)
                 .queryParam("name", name)
-                .queryParam("keepFromSource","all")
                 .queryParam("key", System.getenv("trl_key"))
                 .queryParam("token", System.getenv("trl_token"));
-        return requestSpecification.request(Method.POST, "/cards/" + idCard +"/checklists").as(Checklist.class);
+        return requestSpecification.request(Method.POST, "/cards/" + idCard + "/checklists").as(Checklist.class);
     }
-
 
 
 }
